+++ android/src/main/kotlin/ru/rustore/flutter_rustore_billing/FlutterRustoreBillingClient.kt
@@ -1,8 +1,18 @@
 package ru.rustore.flutter_rustore_billing
 
 import android.app.Application
-import android.util.Log
-import ru.rustore.flutter_rustore_billing.pigeons.Rustore
+import ru.rustore.flutter_rustore_billing.pigeons.ConfirmPurchaseResponse
+import ru.rustore.flutter_rustore_billing.pigeons.DigitalShopGeneralError
+import ru.rustore.flutter_rustore_billing.pigeons.ProductsResponse
+import ru.rustore.flutter_rustore_billing.pigeons.PurchasesResponse
+import ru.rustore.flutter_rustore_billing.pigeons.RustoreBilling
+import ru.rustore.flutter_rustore_billing.pigeons.Subscription
+import ru.rustore.flutter_rustore_billing.pigeons.Product as RustoreProduct
+import ru.rustore.flutter_rustore_billing.pigeons.Purchase as RustorePurchase
+import ru.rustore.flutter_rustore_billing.pigeons.PaymentResult as RustorePaymentResult
+import ru.rustore.flutter_rustore_billing.pigeons.InvalidPurchase as RustoreInvalidPurchase
+import ru.rustore.flutter_rustore_billing.pigeons.SuccessPurchase as RustoreSuccessPurchase
+import ru.rustore.flutter_rustore_billing.pigeons.SubscriptionPeriod as RustoreSubscriptionPeriod
 import ru.rustore.sdk.billingclient.RuStoreBillingClient
 import ru.rustore.sdk.billingclient.RuStoreBillingClientFactory
 import ru.rustore.sdk.billingclient.model.product.Product
@@ -13,10 +23,10 @@ import ru.rustore.sdk.core.config.SdkType
 import ru.rustore.sdk.core.feature.model.FeatureAvailabilityResult
 import ru.rustore.sdk.core.tasks.OnCompleteListener
 
-class FlutterRustoreBillingClient(private val app: Application) : Rustore.RustoreBilling {
+class FlutterRustoreBillingClient(private val app: Application) : RustoreBilling {
     private lateinit var client: RuStoreBillingClient
 
-    override fun initialize(id: String, prefix: String, result: Rustore.Result<String>?) {
+    override fun initialize(id: String, prefix: String, debug: Boolean, callback: (Result<String>) -> Unit) {
         client = RuStoreBillingClientFactory.create(
             context = app,
             consoleApplicationId = id,
@@ -25,194 +35,205 @@ class FlutterRustoreBillingClient(private val app: Application) : Rustore.Rustor
             internalConfig = mapOf("type" to SdkType.FLUTTER)
         )
 
-        result?.success("")
+        callback(Result.success(""))
     }
 
-    override fun available(result: Rustore.Result<Boolean>?) {
+    override fun available(callback: (Result<Boolean>) -> Unit) {
         client.purchases.checkPurchasesAvailability()
             .addOnCompleteListener(object : OnCompleteListener<FeatureAvailabilityResult> {
                 override fun onFailure(throwable: Throwable) {
-                    result?.error(throwable)
+                    callback(Result.failure(throwable))
                 }
 
                 override fun onSuccess(value: FeatureAvailabilityResult) {
                     when (value) {
                         is FeatureAvailabilityResult.Available -> {
-                            result?.success(true)
+                            callback(Result.success(true))
                         }
+
                         is FeatureAvailabilityResult.Unavailable -> {
-                            result?.success(false)
+                            callback(Result.success(false))
                         }
                     }
                 }
             })
     }
 
-    override fun products(
-        ids: MutableList<String>,
-        out: Rustore.Result<Rustore.ProductsResponse>?
-    ) {
-        client.products.getProducts(productIds = ids.toList())
+    override fun products(ids: List<String?>, callback: (Result<ProductsResponse>) -> Unit) {
+        client.products.getProducts(productIds = ids as List<String>)
             .addOnCompleteListener(object : OnCompleteListener<List<Product>> {
                 override fun onFailure(throwable: Throwable) {
-                    out?.error(throwable)
+                    callback(Result.failure(throwable))
                 }
 
                 override fun onSuccess(result: List<Product>) {
-                    val response = Rustore.ProductsResponse.Builder()
-
-                    val products = mutableListOf<Rustore.Product>()
+                    val products = mutableListOf<RustoreProduct>()
 
                     for (item in result) {
-                        val product = Rustore.Product.Builder()
-                            .setProductId(item.productId)
-                            .setProductType(item.productType?.toString())
-                            .setProductStatus(item.productStatus.toString())
-                            .setPriceLabel(item.priceLabel)
-                            .setPrice(item.price?.toLong())
-                            .setCurrency(item.currency)
-                            .setLanguage(item.language)
-                            .setTitle(item.title)
-                            .setDescription(item.description)
-                            .setImageUrl(item.imageUrl.toString())
-                            .setPromoImageUrl(item.promoImageUrl.toString())
-
+                        var subscription: Subscription? = null
 
                         if (item.subscription != null) {
-                            val subscription = Rustore.Subscription.Builder()
-
-                            subscription.setSubscriptionPeriod(period(item.subscription?.subscriptionPeriod))
-                            subscription.setFreeTrialPeriod(period(item.subscription?.freeTrialPeriod))
-                            subscription.setGracePeriod(period(item.subscription?.gracePeriod))
-                            subscription.setIntroductoryPricePeriod(period(item.subscription?.introductoryPricePeriod))
-                            subscription.setIntroductoryPrice(item.subscription?.introductoryPrice)
-                            subscription.setIntroductoryPriceAmount(item.subscription?.introductoryPriceAmount)
-
-                            product.setSubscription(subscription.build())
+                            subscription = Subscription(
+                                subscriptionPeriod = period(item.subscription?.subscriptionPeriod),
+                                freeTrialPeriod = period(item.subscription?.freeTrialPeriod),
+                                gracePeriod = period(item.subscription?.gracePeriod),
+                                introductoryPricePeriod = period(item.subscription?.introductoryPricePeriod),
+                                introductoryPrice = item.subscription?.introductoryPrice,
+                                introductoryPriceAmount = item.subscription?.introductoryPriceAmount
+                            )
                         }
 
-                        products.add(product.build())
+                        val product = RustoreProduct(
+                            productId = item.productId,
+                            productType = item.productType?.toString(),
+                            productStatus = item.productStatus.toString(),
+                            priceLabel = item.priceLabel,
+                            price = item.price?.toLong(),
+                            currency = item.currency,
+                            language = item.language,
+                            title = item.title,
+                            description = item.description,
+                            imageUrl = item.imageUrl.toString(),
+                            promoImageUrl = item.promoImageUrl.toString(),
+                            subscription = subscription,
+                        )
+
+                        products.add(product)
                     }
 
-                    val errors = mutableListOf<Rustore.DigitalShopGeneralError>()
+                    val errors = mutableListOf<DigitalShopGeneralError>()
+                    val response = ProductsResponse(
+                        code = 200,
+                        errors = errors,
+                        products = products,
+                    )
 
-                    response.setProducts(products)
-                    response.setErrors(errors)
-
-                    out?.success(response.build())
+                    callback(Result.success(response))
                 }
             })
     }
 
-    override fun purchase(id: String, out: Rustore.Result<Rustore.PaymentResult>?) {
+    override fun purchase(id: String, callback: (Result<RustorePaymentResult>) -> Unit) {
         client.purchases.purchaseProduct(
             productId = id
         ).addOnCompleteListener(object : OnCompleteListener<PaymentResult> {
             override fun onFailure(throwable: Throwable) {
-                out?.error(throwable)
+                callback(Result.failure(throwable))
             }
 
             override fun onSuccess(result: PaymentResult) {
-                val response = Rustore.PaymentResult.Builder()
+
+                var invalidPurchase: RustoreInvalidPurchase? = null
+                var successPurchase: RustoreSuccessPurchase? = null
 
                 when (result) {
                     is PaymentResult.Cancelled -> {
-                        out?.error(Throwable(message = result.toString()))
+
+                        callback(Result.failure(Throwable(message = result.toString())))
                         return
                     }
+
                     is PaymentResult.Failure -> {
-                        val purchase = Rustore.InvalidPurchase.Builder()
-                        purchase.setPurchaseId(result.purchaseId)
-                        purchase.setInvoiceId(result.invoiceId)
-                        purchase.setOrderId(result.orderId)
-                        purchase.setQuantity(result.quantity?.toLong())
-                        purchase.setProductId(result.productId)
-                        purchase.setErrorCode(result.errorCode?.toLong())
-
-                        response.setInvalidPurchase(purchase.build())
+                        invalidPurchase = RustoreInvalidPurchase(
+                            purchaseId = result.purchaseId,
+                            invoiceId = result.invoiceId,
+                            orderId = result.orderId,
+                            quantity = result.quantity?.toLong(),
+                            productId = result.productId,
+                            errorCode = result.errorCode?.toLong(),
+                        )
                     }
+
                     is PaymentResult.Success -> {
-                        val purchase = Rustore.SuccessPurchase.Builder()
-                        purchase.setOrderId(result.orderId)
-                        purchase.setPurchaseId(result.purchaseId)
-                        purchase.setProductId(result.productId)
-                        purchase.setInvoiceId(result.invoiceId)
-                        purchase.setSubscriptionToken(result.subscriptionToken)
-
-
-                        response.setSuccessPurchase(purchase.build())
+                        successPurchase = RustoreSuccessPurchase(
+                            purchaseId = result.purchaseId,
+                            invoiceId = result.invoiceId,
+                            orderId = result.orderId,
+                            productId = result.productId,
+                            subscriptionToken = result.subscriptionToken,
+                        )
                     }
 
-                    is PaymentResult.InvalidPaymentState -> TODO()
+                    is PaymentResult.InvalidPaymentState -> {
+                        callback(Result.failure(Throwable(message = result.toString())))
+                        return
+                    }
                 }
 
-                out?.success(response.build())
+                val response = RustorePaymentResult(
+                    invalidPurchase = invalidPurchase,
+                    successPurchase = successPurchase,
+                )
+
+                callback(Result.success(response))
             }
         })
     }
 
-    override fun purchases(out: Rustore.Result<Rustore.PurchasesResponse>?) {
+    override fun purchases(callback: (Result<PurchasesResponse>) -> Unit) {
         client.purchases.getPurchases()
             .addOnCompleteListener(object : OnCompleteListener<List<Purchase>> {
                 override fun onFailure(throwable: Throwable) {
-                    out?.error(throwable)
+                    callback(Result.failure(throwable))
                 }
 
                 override fun onSuccess(result: List<Purchase>) {
-                    val response = Rustore.PurchasesResponse.Builder()
-                    response.setCode(200)
-
-                    val purchases = mutableListOf<Rustore.Purchase>()
+                    val purchases = mutableListOf<RustorePurchase>()
 
                     for (item in result) {
-                        val purchase = Rustore.Purchase.Builder()
-                            .setPurchaseId(item.purchaseId)
-                            .setProductId(item.productId)
-                            .setDescription(item.description)
-                            .setLanguage(item.language)
-                            .setPurchaseTime(item.purchaseTime.toString())
-                            .setOrderId(item.orderId)
-                            .setAmountLabel(item.amountLabel)
-                            .setAmount(item.amount?.toLong())
-                            .setCurrency(item.currency)
-                            .setQuantity(item.quantity?.toLong())
-                            .setPurchaseState(item.purchaseState.toString())
-                            .setDeveloperPayload(item.developerPayload)
-                            .setInvoiceId(item.invoiceId)
-                            .setSubscriptionToken(item.subscriptionToken)
+                        val purchase = RustorePurchase(
+                            purchaseId = item.purchaseId,
+                            productId = item.productId,
+                            description = item.description,
+                            language = item.language,
+                            purchaseTime = item.purchaseTime.toString(),
+                            orderId = item.orderId,
+                            amountLabel = item.amountLabel,
+                            amount = item.amount?.toLong(),
+                            currency = item.currency,
+                            quantity = item.quantity?.toLong(),
+                            purchaseState = item.purchaseState.toString(),
+                            developerPayload = item.developerPayload,
+                            invoiceId = item.invoiceId,
+                            subscriptionToken = item.subscriptionToken,
+                        )
 
-                        purchases.add(purchase.build())
+                        purchases.add(purchase)
                     }
 
-                    val errors = mutableListOf<Rustore.DigitalShopGeneralError>()
+                    val errors = mutableListOf<DigitalShopGeneralError>()
+                    val response = PurchasesResponse(
+                        code = 200,
+                        purchases = purchases,
+                        errors = errors,
+                    )
 
-                    response.setErrors(errors)
-                    response.setPurchases(purchases)
-
-                    out?.success(response.build())
+                    callback(Result.success(response))
                 }
             })
     }
 
-    override fun confirm(id: String, out: Rustore.Result<Rustore.ConfirmPurchaseResponse>?) {
+    override fun confirm(id: String, callback: (Result<ConfirmPurchaseResponse>) -> Unit) {
         client.purchases.confirmPurchase(purchaseId = id)
             .addOnSuccessListener {
-                val response = Rustore.ConfirmPurchaseResponse.Builder()
-                response.setSuccess(true)
+                val response = ConfirmPurchaseResponse(
+                    success = true
+                )
 
-                out?.success(response.build())
+                callback(Result.success(response))
             }
             .addOnFailureListener {
-                out?.error(it)
+                callback(Result.failure(it))
             }
     }
 
-    fun period(sub: SubscriptionPeriod?): Rustore.SubscriptionPeriod {
-        val period = Rustore.SubscriptionPeriod.Builder()
-        period.setDays(sub?.days?.toLong() ?: 0)
-        period.setMonths(sub?.months?.toLong() ?: 0)
-        period.setYears(sub?.years?.toLong() ?: 0)
+    fun period(sub: SubscriptionPeriod?): RustoreSubscriptionPeriod {
+        val period = RustoreSubscriptionPeriod(
+            days = sub?.days?.toLong() ?: 0,
+            months = sub?.months?.toLong() ?: 0,
+            years = sub?.years?.toLong() ?: 0,
+        )
 
-        return period.build()
+        return period
     }
 }